server/requirements.txt
Flask==3.0.2
jsonschema==4.21.1
pymongo==4.6.1
python_dateutil==2.8.2
